requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Pillow==10.1.0
pandas==2.1.4
scikit-learn==1.3.2
werkzeug==3.0.1
python-dateutil==2.8.2
face_recognition==1.3.0
Flask-Login==0.6.3
